rust: structure changed ui
